# api/openapi.yaml
openapi: 3.0.3
info:
  title: Blog API
  version: 0.1.0
paths:
  /posts:
    get:
      operationId: listPosts
      responses:
        "200":
          description: list of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPost'
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
  /posts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPost
      responses:
        "200":
          description: post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "404":
          description: not found
    put:
      operationId: updatePost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPost'
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "404":
          description: not found
    delete:
      operationId: deletePost
      responses:
        "204":
          description: deleted
        "404":
          description: not found
components:
  schemas:
    Post:
      type: object
      required: [id, title, content]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
    NewPost:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string